backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
sqlalchemy-serializer==